+++ services/feed-manager-service/feed-manager-controller/src/main/java/com/thinkbiganalytics/feedmgr/service/FeedManagerMetadataService.java
@@ -30,18 +30,19 @@ import com.thinkbiganalytics.feedmgr.rest.model.RegisteredTemplate;
 import com.thinkbiganalytics.feedmgr.rest.model.UIFeed;
 import com.thinkbiganalytics.feedmgr.rest.model.UserFieldCollection;
 import com.thinkbiganalytics.feedmgr.rest.model.UserProperty;
-import com.thinkbiganalytics.feedmgr.security.FeedsAccessControl;
+import com.thinkbiganalytics.feedmgr.security.FeedServicesAccessControl;
 import com.thinkbiganalytics.feedmgr.service.category.FeedManagerCategoryService;
 import com.thinkbiganalytics.feedmgr.service.feed.FeedManagerFeedService;
 import com.thinkbiganalytics.feedmgr.service.feed.FeedModelTransform;
 import com.thinkbiganalytics.feedmgr.service.template.FeedManagerTemplateService;
 import com.thinkbiganalytics.metadata.api.MetadataAccess;
+import com.thinkbiganalytics.metadata.api.category.security.CategoryAccessControl;
 import com.thinkbiganalytics.metadata.api.event.MetadataEventListener;
 import com.thinkbiganalytics.metadata.api.event.MetadataEventService;
 import com.thinkbiganalytics.metadata.api.event.feed.CleanupTriggerEvent;
 import com.thinkbiganalytics.metadata.api.event.feed.FeedOperationStatusEvent;
 import com.thinkbiganalytics.metadata.api.feed.Feed;
-import com.thinkbiganalytics.metadata.api.feedmgr.feed.FeedManagerFeed;
+import com.thinkbiganalytics.metadata.api.feed.security.FeedAccessControl;
 import com.thinkbiganalytics.metadata.api.op.FeedOperation;
 import com.thinkbiganalytics.nifi.rest.client.LegacyNifiRestClient;
 import com.thinkbiganalytics.nifi.rest.client.NiFiComponentState;
@@ -49,6 +50,7 @@ import com.thinkbiganalytics.nifi.rest.client.NiFiRestClient;
 import com.thinkbiganalytics.nifi.rest.model.NifiProperty;
 import com.thinkbiganalytics.nifi.rest.support.NifiProcessUtil;
 import com.thinkbiganalytics.security.AccessController;
+import com.thinkbiganalytics.security.action.Action;
 
 import org.apache.nifi.web.api.dto.ConnectionDTO;
 import org.apache.nifi.web.api.dto.ProcessGroupDTO;
@@ -56,6 +58,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 
 import java.util.Collection;
 import java.util.List;
@@ -66,6 +69,7 @@ import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.inject.Inject;
+import javax.ws.rs.NotFoundException;
 
 /**
  * Provides access to category, feed, and template metadata stored in the metadata store.
@@ -73,6 +77,12 @@ import javax.inject.Inject;
 public class FeedManagerMetadataService implements MetadataService {
 
     private static final Logger log = LoggerFactory.getLogger(FeedManagerMetadataService.class);
+
+    @Value("${kylo.feed.mgr.cleanup.timeout:60000}")
+    private long cleanupTimeout;
+
+    @Value("${kylo.feed.mgr.cleanup.delay:300}")
+    private long cleanupDelay;
 
     @Inject
     FeedManagerCategoryService categoryProvider;
@@ -109,6 +119,13 @@ public class FeedManagerMetadataService implements MetadataService {
      */
     @Inject
     private NiFiRestClient nifiClient;
+    
+    
+    @Override
+    public boolean checkFeedPermission(String id, Action action, Action... more) {
+            return feedProvider.checkFeedPermission(id, action, more);
+
+    }
 
     @Override
     public RegisteredTemplate registerTemplate(RegisteredTemplate registeredTemplate) {
@@ -120,31 +137,6 @@ public class FeedManagerMetadataService implements MetadataService {
         return templateProvider.getTemplateProperties(templateId);
     }
 
-    @Override
-    public RegisteredTemplate getRegisteredTemplate(String templateId) {
-        return templateProvider.getRegisteredTemplate(templateId);
-    }
-
-    @Override
-    public RegisteredTemplate getRegisteredTemplateByName(String templateName) {
-        return templateProvider.getRegisteredTemplateByName(templateName);
-    }
-
-    @Override
-    //@Transactional(transactionManager = "metadataTransactionManager")
-    public RegisteredTemplate getRegisteredTemplateWithAllProperties(final String templateId, String templateName) {
-        return metadataAccess.commit(() -> {
-            return templateProvider.getRegisteredTemplateWithAllProperties(templateId, templateName);
-        });
-    }
-
-    @Override
-    public RegisteredTemplate getRegisteredTemplateForNifiProperties(final String nifiTemplateId, final String nifiTemplateName) {
-        return metadataAccess.commit(() -> {
-            return templateProvider.getRegisteredTemplateForNifiProperties(nifiTemplateId, nifiTemplateName);
-        });
-    }
-
     public void deleteRegisteredTemplate(String templateId) {
         templateProvider.deleteRegisteredTemplate(templateId);
     }
@@ -152,6 +144,12 @@ public class FeedManagerMetadataService implements MetadataService {
     @Override
     public List<RegisteredTemplate> getRegisteredTemplates() {
         return templateProvider.getRegisteredTemplates();
+    }
+
+
+    @Override
+    public RegisteredTemplate findRegisteredTemplateByName(String templateName) {
+        return templateProvider.findRegisteredTemplateByName(templateName);
     }
 
     @Override
@@ -169,15 +167,12 @@ public class FeedManagerMetadataService implements MetadataService {
 
     }
 
-    @Override
-    public void saveFeed(FeedMetadata feed) {
-        feedProvider.saveFeed(feed);
-    }
 
     @Override
     public void deleteFeed(@Nonnull final String feedId) {
         // First check if this should be allowed.
-        this.accessController.checkPermission(AccessController.SERVICES, FeedsAccessControl.ADMIN_FEEDS);
+        accessController.checkPermission(AccessController.SERVICES, FeedServicesAccessControl.ADMIN_FEEDS);
+        feedProvider.checkFeedPermission(feedId, FeedAccessControl.DELETE);
 
         // Step 1: Fetch feed metadata
         final FeedMetadata feed = feedProvider.getFeedById(feedId);
@@ -185,16 +180,19 @@ public class FeedManagerMetadataService implements MetadataService {
             throw new IllegalArgumentException("Unknown feed: " + feedId);
         }
 
-        // Step 2: Check for dependent feeds
+        // Step 2: Check category permissions
+        categoryProvider.checkCategoryPermission(feed.getCategoryId(), CategoryAccessControl.CREATE_FEED);
+
+        // Step 3: Check for dependent feeds
         if (feed.getUsedByFeeds() != null && !feed.getUsedByFeeds().isEmpty()) {
             final List<String> systemNames = feed.getUsedByFeeds().stream().map(FeedSummary::getCategoryAndFeedSystemName).collect(Collectors.toList());
             throw new IllegalStateException("Feed is referenced by " + feed.getUsedByFeeds().size() + " other feeds: " + systemNames);
         }
 
-        // Step 3: Delete hadoop authorization security policies if they exists
+        // Step 4: Delete hadoop authorization security policies if they exists
         if (hadoopAuthorizationService != null) {
             metadataAccess.read(() -> {
-                FeedManagerFeed domainFeed = feedModelTransform.feedToDomain(feed);
+                Feed domainFeed = feedModelTransform.feedToDomain(feed);
                 String hdfsPaths = (String) domainFeed.getProperties().get(HadoopAuthorizationService.REGISTRATION_HDFS_FOLDERS);
 
                 hadoopAuthorizationService.deleteHivePolicy(feed.getSystemCategoryName(), feed.getSystemFeedName());
@@ -203,7 +201,7 @@ public class FeedManagerMetadataService implements MetadataService {
 
         }
 
-        // Step 4: Enable NiFi cleanup flow
+        // Step 5: Enable NiFi cleanup flow
         boolean needsCleanup = false;
         final ProcessGroupDTO feedProcessGroup;
         final ProcessGroupDTO categoryProcessGroup = nifiRestClient.getProcessGroupByName("root", feed.getSystemCategoryName(), false, true);
@@ -215,11 +213,11 @@ public class FeedManagerMetadataService implements MetadataService {
             }
         }
 
-        // Step 5: Run NiFi cleanup flow
+        // Step 6: Run NiFi cleanup flow
         if (needsCleanup) {
             // Wait for input processor to start
             try {
-                Thread.sleep(300);
+                Thread.sleep(cleanupDelay);
             } catch (InterruptedException e) {
                 // ignored
             }
@@ -227,7 +225,7 @@ public class FeedManagerMetadataService implements MetadataService {
             cleanupFeed(feed);
         }
 
-        // Step 6: Remove feed from NiFi
+        // Step 7: Remove feed from NiFi
         if (categoryProcessGroup != null) {
             final Set<ConnectionDTO> connections = categoryProcessGroup.getContents().getConnections();
             for (ProcessGroupDTO processGroup : NifiProcessUtil.findProcessGroupsByFeedName(categoryProcessGroup.getContents().getProcessGroups(), feed.getSystemFeedName())) {
@@ -235,7 +233,7 @@ public class FeedManagerMetadataService implements MetadataService {
             }
         }
 
-        // Step 7: Delete database entries
+        // Step 8: Delete database entries
         feedProvider.deleteFeed(feedId);
 
     }
@@ -273,7 +271,15 @@ public class FeedManagerMetadataService implements MetadataService {
 
     public FeedSummary enableFeed(String feedId) {
         return metadataAccess.commit(() -> {
+            this.accessController.checkPermission(AccessController.SERVICES, FeedServicesAccessControl.EDIT_FEEDS);
+
             FeedMetadata feedMetadata = feedProvider.getFeedById(feedId);
+
+            if (feedMetadata == null) {
+                //feed will not be found when user is allowed to export feeds but has no entity access to feed with feed id
+                throw new NotFoundException("Feed not found for id " + feedId);
+            }
+
             if (!feedMetadata.getState().equals(Feed.State.ENABLED.name())) {
                 FeedSummary feedSummary = feedProvider.enableFeed(feedId);
 
@@ -291,7 +297,14 @@ public class FeedManagerMetadataService implements MetadataService {
 
     public FeedSummary disableFeed(final String feedId) {
         return metadataAccess.commit(() -> {
+            this.accessController.checkPermission(AccessController.SERVICES, FeedServicesAccessControl.EDIT_FEEDS);
+
             FeedMetadata feedMetadata = feedProvider.getFeedById(feedId);
+
+            if (feedMetadata == null) {
+                throw new NotFoundException("Feed not found for id " + feedId);
+            }
+
             if (!feedMetadata.getState().equals(Feed.State.DISABLED.name())) {
                 FeedSummary feedSummary = feedProvider.disableFeed(feedId);
                 boolean updatedNifi = updateNifiFeedRunningStatus(feedSummary, Feed.State.DISABLED);
@@ -351,6 +364,11 @@ public class FeedManagerMetadataService implements MetadataService {
     }
 
     @Override
+    public FeedCategory getCategoryById(String categoryId) {
+        return categoryProvider.getCategoryById(categoryId);
+    }
+
+    @Override
     public void saveCategory(FeedCategory category) {
         categoryProvider.saveCategory(category);
     }
@@ -379,7 +397,7 @@ public class FeedManagerMetadataService implements MetadataService {
             eventService.notify(new CleanupTriggerEvent(feedProvider.resolveFeed(feed.getId())));
 
             // Wait for completion
-            long remaining = 60000L;
+            long remaining = cleanupTimeout;
             while (remaining > 0 && (listener.getState() == null || listener.getState() == FeedOperation.State.STARTED)) {
                 final long start = System.currentTimeMillis();
                 try {
